MVP acabado, falta css

+++ src/pages/AnimalDetails.jsx
@@ -2,42 +2,50 @@ import React from "react";
 import { Link, useParams, useNavigate } from "react-router-dom";
 import { useState, useEffect } from "react";
 import service from "../services/config.services";
+import ClipLoader from "react-spinners/ClipLoader";
+import { AuthContext } from "../context/auth.context";
 
 function AnimalDetails() {
+  const {userRole} = useContext(AuthContext)
+
   const parametrosDinamicos = useParams();
   const navigate = useNavigate();
 
   const [animal, setAnimal] = useState(null);
-
-  useEffect(() => {
-    const getData = async () => {
-      try {
-        const response = await service.get(
-          `/animals/${parametrosDinamicos.animalId}`
-        );
-        console.log(response.data)
-        setAnimal(response.data);
-      } catch (error) {
-        console.log(error);
-      }
-    };
-
-    getData();
-  }, [parametrosDinamicos.animalId]);
-
-  const handleDelete = async () => {
+  const getData = async () => {
     try {
-      await service.delete(`/animals/${parametrosDinamicos.animalId}`);
-      navigate(`/`);
+      const response = await service.get(
+        `/animals/${parametrosDinamicos.animalId}`
+      );
+      console.log(response.data);
+      setAnimal(response.data);
     } catch (error) {
       console.log(error);
     }
   };
 
+  useEffect(() => {
+    getData();
+  }, [parametrosDinamicos.animalId]);
+
+  const handleDelete = async () => {
+    let text = "Are you sure you want to delete it?";
+    if (confirm(text) == true) {
+      try {
+        await service.delete(`/animals/${parametrosDinamicos.animalId}`);
+        navigate(`/`);
+      } catch (error) {
+        console.log(error);
+      }
+    }
+  };
+
   const handleInterested = async () => {
     try {
-      await service.patch(`/animals/${parametrosDinamicos.animalId}/interested`);
-      navigate(`/`);
+      await service.patch(
+        `/animals/${parametrosDinamicos.animalId}/interested`
+      );
+      getData();
     } catch (error) {
       console.log(error);
     }
@@ -53,6 +61,7 @@ function AnimalDetails() {
         }}
       >
         <h3>Buscando data del animal...</h3>
+        <ClipLoader color="green" />
       </div>
     );
   }
@@ -68,22 +77,24 @@ function AnimalDetails() {
       <h3>{animal.age} y/o</h3>
       <h3>{animal.race}</h3>
       <p>{animal.description}</p>
-      {animal.interested.map((eachInterested)=>{
-        return(
-        <p>{eachInterested.name}</p>
-        )
+      {animal.interested.map((eachInterested) => {
+        return <p>{eachInterested.name}</p>;
       })}
 
-      <button id="btnParticipate" onClick={handleInterested}>
-        Apply for adoption
-      </button>
-
-      <button className="btnDelete" onClick={handleDelete}>
-        Delete
-      </button>
-      <Link to={`/animals/edit/${parametrosDinamicos.animalId}`}>
-        <button className="btnEdit">Edit</button>
-      </Link>
+      {userRole === "admin" ? (
+        <>
+          <button className="btnDelete" onClick={handleDelete}>
+            Delete
+          </button>
+          <Link to={`/animals/edit/${parametrosDinamicos.animalId}`}>
+            <button className="btnEdit">Edit</button>
+          </Link>
+        </>
+      ) : (
+        <button id="btnParticipate" onClick={handleInterested}>
+          Apply for adoption
+        </button>
+      )}
     </div>
   );
 }
